requirements.txt
anyio==3.3.4
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.4
colorama==0.4.4
cycler==0.11.0
dash==2.3.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.3
filelock==3.6.0
Flask==2.0.3
Flask-Compress==1.11
gunicorn==20.1.0 
idna==3.3
influxdb==5.3.1
influxdb-client==1.26.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
itsdangerous==2.1.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
nbclassic==0.3.4
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.1
plotly==5.6.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
pycparser==2.21
Pygments==2.10.0
pyorbital==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
Rx==3.2.0
scikit-learn==1.0.1
scipy==1.7.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
virtualenv==20.13.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.3
